refactor: handle auth type selection, clean up comments in code files
